Unfinished Additions to Markers
It now grabs the last known location, centers to it, and drops a red
marker there.

Then it puts together an array of 6 other markers, 3 Blue markers with
the title "Internet", and 3 Yellow markers with the title "Electricity".
Need to work on how it will be stored and pulled from SQL before much
more is done here.