SQLite database implemented
The MapsActivity will now check for the existence of existing tables,
and if they do not exist, will create the required tables. The
MapsActivity will then parse the tables, and create Markers from the
data.

For now, on launch insertSampleData() is run which will drop all data,
then re-add ensuring even testing data on all platforms at launch. It is
the last method in the file "MapsActivity.java" and is an excellent
resource for learning how to navigate the SQLite database.