Migrated database code into a class
All code for interfacing with SQLite has been moved into a
class,outageDatabase that can be instantiated with the argument
"(this)".

The database is split into one table per utility type. This way it will
be easy to filter utility types. As such, there are methods added for:
getInternetReports
getElectricityReports
...
getInternetMarkers
getElectricitiyMarkers
...
insertReport
insertSampleData (Temporary for wiping clean at first launch).

The Maps Activity successfully pulls from the database through this
manner, and I have added a "test insert" that will run automatically
when you open the "createReportActivity".

For the map to get updated to see the inserted marker, I currently am
forcing the map to wipe all markers and re-fetch them on each resume.
This bogs it down a fait bit and should be revised to only run on
insert, and on a requested "Refresh".