Added userID support
I made it where it will wipe the db and put sample data only on the very
first launch. So if you need to start over, uninstall the app, then push
it back to the device.

On the first launch it will generate a random number between 11111 and
99999. It is then saved in SharedPreferences under the tag "userID".

The tab "Reports" has been relabeled "My Reports" and will only show
those based on the user's ID.

The createReport page has been edited to put the stored userID into the
new report to support the other features.